Generate the Yesod scaffolded site
    cabal install yesod-platform
    yesod init
    cd rivierarb
    yesod devel

The safest way to install Yesod is through the yesod-platform package:
it references a consistent (and tested) set of the Yesod dependencies.
That should avoid dependency hell.  The even safer way to install it
would be through cabal-dev.  It's actually the one I used for the live
demo, but it's completely out of scope for the presentation.  Suffice
to say: I had not encountered Haskell dependency madness before
probing the web frameworks, but installing Snap and Yesod in parallel
on the same user account wasn't the smoothest experience of my life.

`yesid init` prompts us for some project information.  The only
relevant one for the demo was the project name; for the rest we picked
the defaults.  `yesod devel` launches the scaffolded server in
development mode, to be targetted on http://localhost:3000/