Start taking possession of the scaffold
Our first commit is mainly there to validate we actually own the code
that powers the development server.  We change the homepage template;
the server automatically restarts; we refresh the page in the brower.
Lo and behold: our change made it to HTTP!

That first commit also serves as a first contact with Hamlet, the
templating language Yesod uses.  The main point to notice for now is
that indentation replaces closing tags, but there is a lot more to
come.

Our commit also happened to create the first dead (as in 404) link in
our site.  Let's fix that.