Create a URI; repair a 404
The first step in creating our blog's landing page is to set up a
route for it.  Routes in the Yesod scaffold are defined in
`config/routes`.  Mocking the route for the index page, we define URI
`/blog`, of representation type `BlogR`, answering to sole method
`GET`.

Yesod notices the change and tries to recompile and relaunch the
development server.  But it fails on a type error.  Our change to the
route file has resulted in the `BlogR` type being insufficiently
defined.  Haskell catches that at compile time.

Again mocking the index page defined in `Handler/Home.hs` (and staying
in that file; Cabal is out of scope for this live session), we define
a simple handler for our page.  It references a non-existing template
file.  This results in another compile-time error until we create a
stub for it.