Typesafe links
Our blog is a single static page, and already we've got redundancy:
its URI is defined as text in the route file, in addition to being
linked from the homepage.  If we ever changed the route, our homepage
link would 404 again.

To prevent this, we convert the quoted link to a typesafe one using
Hamlet `@`-syntax.  Now our referenced URI is little more than a
typename.  We can check immediately that:

  * we can move both the blog page address and its homepage link by
    simply changing the route file.

  * the compiler won't let us put anything wrong in place of that
    typesafe link.