Hamlet syntax and Haskell records
We create a simple Haskell record type to represent a blog post, and a
list variable to store them.  We can now refer to it from our blog
template, and demonstrate the more elaborate syntax for conditionals
(same as Haskell conditionals), loops (same as Haskell list
comprehensions) and interpolations.

We notice that the template is really a part of the code now: it
references variables reachable from the module where the template is
included only.  This is made possible by Yesod's use of Template
Haskell.

Now is also the good time to notice the typesafe effects of Hamlet
interpolation: try injecting a `<script>` element with
carefully-crafted `Post` entries.  Yesod protects against HTML
injection automatically: it recognizes that the interpolated post
contents (or title) is of Haskell type `Text`, and escapes whatever
needs be.