Persistence: reading
At last we replace our global `posts` variable with one local to the
blog handler, that fetches the information from the database.  The two
empty lists in argument to `selectList` are typesafe filters (WHERE
clauses) and options (the ORDER BY sort of thing) respectively.

This change has a slight effect on the type the posts are refered to
as in the template, so we have to adjust for that before the compiler
lets us through.

We can now create (and modify, and delete) blog entries with an SQL
console in a separate window.  We'll do better in a moment.