Typesafe links with arguments
Right now, all we see from the posts is their title.  Let's create a
specific page to exhibit them to the world!

First, the route.  It takes a type-checked argument of `PostId` type,
defined through the model.  Blog terminology is a bit unfortunate
here: let's not get confused by the fact we're defining a _Blog_ post
handler, expected to answer _HTTP_ GET requests.  We've got plenty of
time to tackle HTTP POSTs later on.

Then, the handler.  Notice it takes an argument—we can guess its type.
Please close your eyes on the partial match I just introduced, it's
for the presentation terseness's sake (and only temporary until next
commit, honest!)

We create a new template for the post and we're done, as far as
compilation success is concerned.  For blog discoverability from the
user's point of view, we still need to link to the posts.  That's our
chance to use a typesafe link with a type-checked argument.