SQL Error handling
Remember that partial match?  With our typesafe URIs, we can't have a
link generated that points to a post id that doesn't exist.  But it's
not fully out of the picture just yet: there's no saying a user won't
display the post list, wait long enough for someone else to erase an
entry, and then try and click.  Or that a particularly malicious one
would try and edit the address bar by hand.  In that case we get a
nasty-looking 500 Internal Error response.

Replacing that `get` with a `get404` solves this elegantly:

  * no more internal error, because

  * no more partial match

  * an actual 404 Not Found page, the responsible HTTP citizen's way
    of doing

I'll gloss over the visual similarity between the scaffolded site's
default 404 and 500 pages.