HTTP POST and write-through persistence
To allow creating blog posts from the web interface, we add an HTML
form to the end of the blog page template.  The relevant POST handler
is declared in the routes file, and defined along with its GET
counterpart.

I concede two partial matches in a single commit (I'll fix, I'll fix!)
But I boast an elegant demonstration of both the persistence layer's
write syntax and a typesafe HTTP redirection.