Typesafe forms: generation
Did you notice, in addition to the partial matches, the clumsy
repetition of the field names as text in both the Hamlet template and
the POST handler?  We'll fix all of that immediately using a Yesod
form.

This syntax is the scariest Haskell we'll see today.  It's
unfortunately way out of scope for this simple live session.  The key
points to keep in mind about this “entity” are:

  * it's supposed to be rendered in templates with `<div>` form
    (alternatives include HTML tables)
  * it handles `Post`-type records
  * it builds them from two `req`uired fields,
  * of type text (`textField`)
  * quoted labels (`"Title"` and `"Content"`)
  * no default value (`Nothing`)

( The key points to forget are: why that letter “a” in `areq`, and
what those `<$>` and `<*>` do.  The search engine keywords are
“applicative functor”, they're definitely not beginner-friendly.  But
it's the shortest way to implement forms that match the model
structure so closely. )

When splicing the form in our templates, we need that `enctype` field
in case the form contains a file-type field, since the actual `<form>`
tag is out of the generator's scope.