Rich HTML editor
For our last commit, we'll upgrade our post write-up textbox to a
full-fledged HTML editor.  (This one needs actual Internet
connectivity, so I couldn't verify it in the plane, but it's fine
now.)  As an exercice, count the number of lines altered in the source
code to implement that.  Keeping in mind the ones in `Foundation.hs`
are common to the whole application.  Also try and inject malicious
`<script>` elements without resorting to SQL.